refactor(files): decompose FileViewer.tsx monolith
Split the 1681-line FileViewer.tsx into focused units, eliminating the
duplicated wide/mobile render switch (822-892 ≈ 968-1031):

- FileContentView.tsx: the single image/media/markdown/html/code/diff
  switch, rendered by both the wide and mobile layouts
- ChangesView.tsx: ChangesView / TreeView / buildTree / countLeaves /
  gitStatusLabel and the changes source/display storage
- useViewHistory.ts: back-navigation stack (history.pushState / popstate
  / Escape / unmount cleanup)
- file-types.ts: shared extension detection + getFileName

FileViewer.tsx is now layout + state orchestration only. Behavior is
unchanged; media now uses the wide (full-height + caption) layout on
mobile too, and the mobile CodeViewer now receives filePath like wide.

Closes #312
